+++ dsp/phase_distortion/double_kink.hpp
@@ -9,13 +9,13 @@ class DoubleKinkAlgo : public PDAlgo {
         DoubleKinkAlgo();
         ~DoubleKinkAlgo() {}
 
-        virtual uint32_t process_phase(uint32_t pha);
+        virtual float process_phase(float pha);
 
     private:
         virtual void update_params_();
         float interp_;
-        uint32_t kink_point_1_, kink_point_2_;
-        uint32_t offset_1_, offset_2_;
+        float kink_point_1_, kink_point_2_;
+        float offset_1_, offset_2_;
 };
 
 #endif  // PHASE_DISTORTION_DOUBLE_KINK_H_
