+++ dsp/phase_distortion/sync.hpp
@@ -8,7 +8,7 @@ class SyncAlgo : public PDAlgo {
         SyncAlgo() {}
         ~SyncAlgo() {}
 
-        virtual uint32_t process_phase(uint32_t pha);
+        virtual float process_phase(float pha);
 
     private:
         virtual void update_params_();
